Update waveshare43.yaml
Added button 5, 6, 7 entities. Edited the date format. Have a gap between day and month that would like to close, but not a bit deal. Added 12-hour time format. Time font was my_font120 which was too large for the display, tried to change size to a little smaller and its currently too small.